requirements.txt
fastapi==0.88.0
google_api_python_client==2.70.0
isodate==0.6.1
python-dotenv==0.21.0
pytube==12.1.0
requests==2.28.1
uvicorn==0.20.0
openai